refactor(blorktools): state to scene
