+++ packages/blorktools/src/asset_debugger/landing-page/dropzone-preview.js
@@ -340,7 +340,7 @@ function addClearButton(container, file) {
                 });
                 
                 // Reset state model file
-                import('../state.js').then(stateModule => {
+                import('../scene/state.js').then(stateModule => {
                     const state = stateModule.getState();
                     state.modelFile = null;
                     state.useCustomModel = false;
