+++ packages/blorktools/src/asset_debugger/panels/asset-panel/asset-panel.js
@@ -3,7 +3,7 @@
  * 
  * This module handles sample sections and their collapsible functionality.
  */
-import { getState } from '../../state';
+import { getState } from '../../scene/state';
 import { createMeshVisibilityPanel } from '../mesh-panel/mesh-panel';
 import { initAtlasPanel, updateAtlasVisualization } from '../atlas-panel/atlas-panel';
 import { initUvPanel, updateUvPanel } from '../uv-panel/uv-panel.js';
