+++ packages/blorktools/src/asset_debugger/panels/atlas-panel/atlas-panel.js
@@ -3,7 +3,7 @@
  * 
  * This module handles texture atlas visualization in the UI.
  */
-import { getState, updateState } from "../../state.js";
+import { getState, updateState } from "../../scene/state.js";
 
 // Track initialization state
 let isInitialized = false;
@@ -74,7 +74,7 @@ function setupAtlasPanelEvents() {
  */
 export function updateAtlasVisualization() {
     // Get state to retrieve texture objects
-    import('../../state.js').then(stateModule => {
+    import('../../scene/state.js').then(stateModule => {
         const state = stateModule.getState();
         
         // Check if we have texture objects
