+++ packages/blorktools/src/asset_debugger/scene/axis-indicator/axis-indicator.js
@@ -1,5 +1,5 @@
 import * as THREE from 'three';
-import { getState } from '../../state';
+import { getState } from '../state';
 
 // Add variables to track axis indicator state
 let axisIndicatorCollapsed = false;
