+++ packages/blorktools/src/asset_debugger/util/rig/bone-util.js
@@ -1,7 +1,7 @@
 import * as THREE from 'three';
 import { rigOptions } from './rig-manager';
 import { getIsDragging } from './drag-util';
-import { getState } from '../../state';
+import { getState } from '../../scene/state';
 // These variables need to be exported so they're available to both modules
 export let bones = [];
 export let boneVisualsGroup = null;
