+++ packages/blorktools/src/asset_debugger/util/ui-manager.js
@@ -3,7 +3,7 @@
  * 
  * This module handles the UI interactions and tab switching.
  */
-import { getState } from '../state.js';
+import { getState } from '../scene/state.js';
 import { updateAtlasVisualization } from '../panels/atlas-panel/atlas-panel.js';
 import { updateUvPanel } from '../panels/uv-panel/uv-panel.js';
